Fix a few issues in the less compiler such as support for ConsString, java.io.File handling, and support for less files using newer syntax
